Allowing a custom file reader be passed into FileDataSource
